feat: ai 관련 엔드포인트 및 서비스 계층구조 완성

+++ ProjecTree-Inference/app/api/schemas/nodes.py
@@ -7,8 +7,8 @@ class NodeCreateRequest(BaseModel):
     workspace_id: int = Field(..., description="워크스페이스 ID")
     candidate_id: int = Field(..., description="후보 노드 ID")
     parent_id: int = Field(..., description="부모 노드 ID")
-    x_pos: int = Field(..., description="노드 X 좌표")
-    y_pos: int = Field(..., description="노드 Y 좌표")
+    x_pos: float = Field(..., description="노드 X 좌표")
+    y_pos: float = Field(..., description="노드 Y 좌표")
 
 
 class NodeCreateResponse(BaseModel):
@@ -16,6 +16,6 @@ class NodeCreateResponse(BaseModel):
     node_id: Optional[int] = Field(None, description="생성된 노드 ID")
     parent_id: Optional[int] = Field(None, description="부모 노드 ID")
     candidate_id: Optional[int] = Field(None, description="후보 노드 ID")
-    x_pos: Optional[int] = Field(None, description="노드 X 좌표")
-    y_pos: Optional[int] = Field(None, description="노드 Y 좌표")
+    x_pos: Optional[float] = Field(None, description="노드 X 좌표")
+    y_pos: Optional[float] = Field(None, description="노드 Y 좌표")
